BUG: 8275. itkShrinkImageFilter does not preserve object physical locations. These changes cause the physical location of the center of the output to be the same as the center of the input. As a side-effect, the origins will not necessarily coincide. This new behaviour is more appropriate since the ShrinkFilter is often used for multi resolution registration. And, arguably, is the behavior one would expect in general. Also, these changes reuqired changes to some tests that expected the origins to coincide. Tests that checked the output of the filter to know inputs also required changes to select the approprite true value. BUG: 8245. itkShrinkImageFilter was not streaming because of an unnecessary EnlargeOutputRequestedRegion.

+++ Testing/Code/BasicFilters/itkShrinkImagePreserveObjectPhysicalLocations.cxx
@@ -23,7 +23,7 @@
 //typedef itk::Image<signed short, 2> TImageType;
 typedef itk::Image<signed short, 2> WImageType;
 //typedef itk::Image<signed short, 2> TImageType;
-typedef itk::Image<float, 2> TImageType;
+typedef itk::Image<double, 2> TImageType;
 //typedef itk::Image<double, 2> TImageType;
 
 //Need to use a Pyramid filter here instead of just downsampling to a 32x32 image
